chore: move constants file
